@@ -7,6 +7,7 @@
     "axios": "^0.15.3",
     "babel": "^6.23.0",
     "babel-core": "^6.23.1",
+    "babel-polyfill": "^6.23.0",
     "gulp": "^3.9.1",
     "live-server": "^1.1.0",
     "react": "^15.3.1",
@@ -21,7 +22,6 @@
     "watchify": "^3.7.0"
   },
   "devDependencies": {
-    "babel-polyfill": "^6.23.0",
     "babel-preset-airbnb": "2.2.3",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.11.1",
@@ -32,8 +32,11 @@
     "chalk": "^1.1.3",
     "del": "^2.2.2",
     "enzyme": "^2.7.1",
-    "eslint": "^3.15.0",
-    "eslint-plugin-react": "^6.9.0",
+    "eslint": "^3.17.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.10.0",
     "gulp-babel-istanbul": "^1.6.0",
     "gulp-cache": "^0.4.5",
     "gulp-clean-css": "^3.0.2",
