@@ -13,7 +13,7 @@ class BodyLayout extends PureComponent {
 }
 
 BodyLayout.propTypes = {
-  children: React.PropTypes.element,
+  children: React.PropTypes.element.isRequired,
 };
 
 export default BodyLayout;
