@@ -4,13 +4,13 @@ export const APP_ERROR = 'APP_ERROR';
 export function appLoadingAction(status) {
   return {
     type: APP_IS_LOADING,
-    payload: status
+    payload: status,
   };
 }
 
 export function appErrorAction(error) {
   return {
     type: APP_ERROR,
-    payload: error
+    payload: error,
   };
 }
