@@ -2,9 +2,9 @@ import React from 'react';
 
 const SpinLoader = () => (
   <div className="c-spin-loader">
-    <div className="bounce1"></div>
-    <div className="bounce2"></div>
-    <div className="bounce3"></div>
+    <div className="bounce1" />
+    <div className="bounce2" />
+    <div className="bounce3" />
   </div>
 );
 
